+++ tests/testthat/test-likelihood.R
@@ -1,4 +1,29 @@
-context("likelihoods")
+
+test_that("likelihood() catches input errors", {
+  x = singleton(1)
+  y = nuclearPed(1)
+  m = marker(x)
+  expect_error(likelihood(x, y), "Invalid input for argument `markers`")
+  expect_error(likelihood(x, y, m), "Invalid input for argument `markers`")
+  expect_error(likelihood(x, y, marker = m), "Invalid input for argument `peelOrder`")
+
+  expect_error(likelihood(x, marker = list(1)), "Invalid input for argument `markers`.")
+  expect_error(likelihood(x, marker = NA), "`markers` is a logical of length 1")
+  expect_error(likelihood(x, marker = 1), "Marker index out of range: 1")
+  expect_error(likelihood(x, marker = "a"), "Unknown marker name: a")
+
+  peds = list(x,y)
+  expect_error(likelihood(peds, m), fixed = T,
+               "`likelihood.list()` requires `markers` to be a vector of marker names or indices.")
+  expect_error(likelihood(peds, list()), fixed = T,
+               "`likelihood.list()` requires `markers` to be a vector of marker names or indices.")
+  expect_error(likelihood(peds, 1), "Marker index out of range: 1")
+
+  peds2 = list(setMarkers(x,marker(x)), y)
+  expect_error(likelihood(peds2), "The pedigree components have different number of markers attached")
+  expect_error(likelihood(peds2,1), "Marker index out of range: 1")
+})
+
 
 test_that("empty markers give likelihood 1", {
   x = nuclearPed(1)
@@ -12,8 +37,8 @@ test_that("empty markers give likelihood 1", {
   x3 = halfCousinPed(0)
   expect_equal(likelihood(x3, marker(x3)), 1)
 
-  #x4 = cousinPed(0, child=T)
-  #expect_equal(likelihood(x4, marker(x4)), 1)
+  x4 = cousinPed(0, child=T)
+  expect_equal(likelihood(x4, marker(x4)), 1)
 })
 
 test_that("nuclear likelihoods are correct, HW-like", {